requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-groq>=0.2.0
langchain-google-genai>=2.0.0
google-generativeai>=0.5.0
# transformers>=4.40.0
# torch>=2.0.0
# Pillow>=10.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
docling>=2.0.0
google-genai